+++ JS/views/components/card/BothCardView.js
