+++ JS/views/components/card/CardFooterView.js
