+++ JS/views/components/card/CardHeaderView.js
