+++ JS/views/components/card/OnlyImgCardView.js
