+++ JS/views/components/card/OnlyTextCardView.js
